+++ R/embrs.local.R
@@ -1,0 +1,102 @@
+###############################
+#currently unexported functions
+###############################
+
+###############################
+#embrs_vehicle
+##############################
+
+#workhorse function for making
+#vehicle/fleet objects
+
+####################
+#to do
+####################
+
+#doc args
+
+embrs_vehicle <-
+  function(name = NULL, x = NULL, ...){
+
+    .obj <- list(...)
+    #print(.obj)
+    if(is.null(x)){
+      stop("nothing to work with!")
+    }
+    obj <- list(args=list(),
+                funs=list()
+    ) # thinking of adding docs as extra case
+    #separate all functions start ef.
+    test <- grep("^ef[.]", tolower(names(.obj)))
+    .ef <- if(length(test)>0) {
+      .obj[test]
+    } else {
+      list()
+    }
+    test <- which(!1:length(.obj) %in% test)
+    .obj <- .obj[test]
+    #x args and funs to output
+    if(is.list(x$args)){
+      obj$args <- modifyList(obj$args, x$args)
+    }
+    if(is.list(x$funs)){
+      obj$funs <- modifyList(obj$funs, x$funs)
+    }
+    #... args and .ef to output
+    if(is.list(.obj)){
+      obj$args <- modifyList(obj$args, .obj)
+    }
+    if(is.list(.ef)){
+      obj$funs <- modifyList(obj$funs, .ef)
+    }
+
+    name <- if(is.null(name)){
+      paste(obj$args$eng.fuel,  obj$args$eng.type, obj$args$veh.type, sep=".")
+    } else { as.character(name) }
+    obj$args$name <- name
+
+    out <- obj
+    class(out) <- "embrs_vehicle"
+    out <- list(out)
+    names(out)[1] <- out[[1]]$args$name
+    out <- list(fleet=out)
+    class(out) <- c("embrs", "fleet")
+    out
+  }
+
+
+#############################
+#route_route
+#############################
+
+#used to build route_objects
+
+####################
+#to do
+####################
+
+#doc args
+
+embrs_route <-
+  function(name = NULL, route.def = NULL, route.source = NULL,
+           route.dist = 1, route.slope = 0, ...){
+    if(is.null(route.def) || is.null(route.source)){
+      stop("[embrs] route_...() needs route.def and route.source [see help]",
+           call. = FALSE)
+    }
+    obj <- list(args=list(name=name,
+                          route.def=route.def,
+                          route.source=route.source,
+                          route.dist=route.dist,
+                          route.slope=route.slope,
+                          ...))
+    if(is.null(obj$args$name)){
+      obj$args$name <- paste(obj$args$route.source,  obj$args$route.def, sep=".")
+    }
+    class(obj) <- "embrs_route"
+    out <- list(obj)
+    names(out)[1] <- out[[1]]$args$name
+    out <- list(routes=out)
+    class(out) <- c("embrs", "routes")
+    out
+  }