+++ R/route.objects.R
@@ -17,37 +17,6 @@
 #' @note trying to streamline these, so likely to be subject to change
 ## #' @references [to do] add embrs and naei references...
 
-#############################
-#testing route_workhorse
-#and route functions built using that...
-#############################
-
-#splatted function
-#' @rdname route.objects
-#' @export
-embrs_route <-
-  function(name = NULL, route.def = NULL, route.source = NULL,
-           route.dist = 1, route.slope = 0, ...){
-    if(is.null(route.def) || is.null(route.source)){
-      stop("[embrs] route_...() needs route.def and route.source [see help]",
-           call. = FALSE)
-    }
-    obj <- list(args=list(name=name,
-                         route.def=route.def,
-                         route.source=route.source,
-                         route.dist=route.dist,
-                         route.slope=route.slope,
-                         ...))
-    if(is.null(obj$args$name)){
-      obj$args$name <- paste(obj$args$route.source,  obj$args$route.def, sep=".")
-    }
-    class(obj) <- "embrs_route"
-    out <- list(obj)
-    names(out)[1] <- out[[1]]$args$name
-    out <- list(routes=out)
-    class(out) <- c("embrs", "routes")
-    out
-  }
 
 #splatted function
 #' @rdname route.objects